Move the cpu-side reduce utility from reduce_common.hpp to test/cpu_reduce_util.hpp
