+++ src/include/miopen/reduce_common.hpp
@@ -27,22 +27,9 @@
 #define GUARD_MIOPEN_REDUCE_COMMON_HPP
 
 #include <half.hpp>
-#include <limits>
-#include <cmath>
-#include <miopen/miopen.h>
-#include <miopen/float_equal.hpp>
-
-#include "bfloat16.hpp"
+#include <miopen/bfloat16.hpp>
 
 namespace reduce {
-
-enum ReductionMethod_t
-{
-    Reduce_DirectThreadWise = 1,
-    Reduce_DirectWarpWise   = 2,
-    Reduce_BlockWise        = 3,
-    Reduce_MultiBlock       = 4
-};
 
 // data type conversion
 template <typename T>
@@ -81,229 +68,6 @@ template <>
 inline bfloat16 type_convert<bfloat16>::operator()<float>(float x) const
 {
     return bfloat16(x);
-};
-
-template <typename compType>
-static inline std::function<void(compType&, compType)> ReduceOpFn(miopenReduceTensorOp_t op_)
-{
-    switch(op_)
-    {
-    case MIOPEN_REDUCE_TENSOR_ADD: return ([&](compType& a_, compType b_) { a_ = a_ + b_; });
-
-    case MIOPEN_REDUCE_TENSOR_MUL: return ([&](compType& a_, compType b_) { a_ = a_ * b_; });
-
-    case MIOPEN_REDUCE_TENSOR_MIN:
-        return ([&](compType& a_, compType b_) {
-            if(a_ > b_)
-                a_ = b_;
-        });
-
-    case MIOPEN_REDUCE_TENSOR_MAX:
-        return ([&](compType& a_, compType b_) {
-            if(a_ < b_)
-                a_ = b_;
-        });
-    }
-
-    return (std::function<void(compType&, compType)>{});
-};
-
-template <typename compType>
-static inline std::function<void(compType&, compType, bool& changed)>
-ReduceOpFn2(miopenReduceTensorOp_t op_)
-{
-    switch(op_)
-    {
-    case MIOPEN_REDUCE_TENSOR_MIN:
-        return ([&](compType& a_, compType b_, bool& changed) {
-            if(a_ > b_)
-            {
-                a_      = b_;
-                changed = true;
-            }
-            else
-                changed = false;
-        });
-
-    case MIOPEN_REDUCE_TENSOR_MAX:
-        return ([&](compType& a_, compType b_, bool& changed) {
-            if(a_ < b_)
-            {
-                a_      = b_;
-                changed = true;
-            }
-            else
-                changed = false;
-        });
-
-    case MIOPEN_REDUCE_TENSOR_ADD:
-    case MIOPEN_REDUCE_TENSOR_MUL: return (std::function<void(compType&, compType, bool&)>{});
-    };
-
-    return (std::function<void(compType&, compType, bool&)>{});
-};
-
-template <typename compType>
-static inline compType ReduceOpZeroVal(miopenReduceTensorOp_t op_)
-{
-    switch(op_)
-    {
-    case MIOPEN_REDUCE_TENSOR_ADD: return (type_convert<compType>{}(0.0));
-
-    case MIOPEN_REDUCE_TENSOR_MUL: return (type_convert<compType>{}(1.0));
-
-    case MIOPEN_REDUCE_TENSOR_MIN: return (std::numeric_limits<compType>::max());
-
-    case MIOPEN_REDUCE_TENSOR_MAX: return (std::numeric_limits<compType>::min());
-    }
-
-    return (type_convert<compType>{}(0.0));
-};
-
-template <>
-inline half_float::half ReduceOpZeroVal<half_float::half>(miopenReduceTensorOp_t op_)
-{
-    switch(op_)
-    {
-    case MIOPEN_REDUCE_TENSOR_ADD: return (type_convert<half_float::half>{}(0.0));
-
-    case MIOPEN_REDUCE_TENSOR_MUL: return (type_convert<half_float::half>{}(1.0));
-
-    case MIOPEN_REDUCE_TENSOR_MIN:
-        return (type_convert<half_float::half>{}(std::numeric_limits<float>::max()));
-
-    case MIOPEN_REDUCE_TENSOR_MAX:
-        return (type_convert<half_float::half>{}(std::numeric_limits<float>::min()));
-    }
-
-    return (type_convert<half_float::half>{}(0.0));
-};
-
-template <typename T>
-static inline bool IsNan(T x)
-{
-    // C++ isnan() is used for float and double
-    return (std::isnan(x));
-};
-
-template <>
-inline bool IsNan<half_float::half>(half_float::half x)
-{
-    return (half_float::isnan(x));
-};
-
-template <typename T>
-static inline bool IsFinite(T x)
-{
-    // C++ isfinite() is used for float and double
-    return (std::isfinite(x));
-};
-
-template <>
-inline bool IsFinite<half_float::half>(half_float::half x)
-{
-    return (half_float::isfinite(x));
-};
-
-struct float_equal_one
-{
-    template <class T>
-    static bool apply(T x)
-    {
-        return std::isfinite(x) and
-               std::nextafter(x, std::numeric_limits<T>::lowest()) <= static_cast<T>(1.0) and
-               std::nextafter(x, std::numeric_limits<T>::max()) >= static_cast<T>(1.0);
-    }
-
-    template <class T>
-    bool operator()(T x)
-    {
-        return (float_equal_one::apply(x));
-    };
-};
-
-struct float_equal_zero
-{
-    template <class T>
-    static bool apply(T x)
-    {
-        return std::isfinite(x) and
-               std::nextafter(x, std::numeric_limits<T>::lowest()) <= static_cast<T>(0.0) and
-               std::nextafter(x, std::numeric_limits<T>::max()) >= static_cast<T>(0.0);
-    }
-
-    template <class T>
-    bool operator()(T x)
-    {
-        return (float_equal_zero::apply(x));
-    };
-};
-
-template <>
-inline bool float_equal_one::apply<half_float::half>(half_float::half x)
-{
-    return half_float::isfinite(x) and x <= type_convert<half_float::half>{}(1.0) and
-           x >= type_convert<half_float::half>{}(1.0);
-};
-
-template <>
-inline bool float_equal_zero::apply<half_float::half>(half_float::half x)
-{
-    return half_float::isfinite(x) and x <= type_convert<half_float::half>{}(0.0) and
-           x >= type_convert<half_float::half>{}(0.0);
-};
-
-template <typename compType>
-static inline void binop_with_nan_check(miopenNanPropagation_t nanOpt,
-                                        std::function<void(compType&, compType)> opReduce,
-                                        compType& accuVal,
-                                        compType currVal)
-{
-    if(nanOpt == MIOPEN_NOT_PROPAGATE_NAN)
-        opReduce(accuVal, currVal);
-    else
-    {
-        if(reduce::IsNan(currVal))
-            accuVal = currVal;
-        else
-            opReduce(accuVal, currVal);
-    };
-};
-
-template <typename compType>
-static inline void binop_with_nan_check2(miopenNanPropagation_t nanOpt,
-                                         std::function<void(compType&, compType, bool&)> opReduce,
-                                         compType& accuVal,
-                                         compType currVal,
-                                         int& accuIndex,
-                                         int currIndex)
-{
-    if(nanOpt == MIOPEN_NOT_PROPAGATE_NAN)
-    {
-        bool changed;
-
-        opReduce(accuVal, currVal, changed);
-
-        if(changed)
-            accuIndex = currIndex;
-    }
-    else
-    {
-        if(reduce::IsNan(currVal))
-        {
-            accuVal   = currVal;
-            accuIndex = currIndex;
-        }
-        else
-        {
-            bool changed;
-
-            opReduce(accuVal, currVal, changed);
-
-            if(changed)
-                accuIndex = currIndex;
-        };
-    };
 };
 
 }; // end of namespace reduce
